requirements.txt
# LangChain Core
langchain==0.3.27
langchain_core==0.3.79

# LangChain Components
langchain_text_splitters==0.3.11
langchain_openai==0.3.35
langchain_google_genai==2.1.12
langchain_postgres==0.0.16
langchain-community==0.3.19

# Utilities
pydantic==2.12.3
typer==0.20.0

# PostgreSQL Drivers
psycopg==3.2.11
psycopg-binary==3.2.11
psycopg-pool==3.2.6

# Environment Management
python-dotenv==1.0.0
